render embed page with theme
