+++ console/components/channel/ChannelView.jsx
@@ -1,20 +1,14 @@
 import React, { PureComponent } from 'react'
 import insertTextAtCursor from 'insert-text-at-cursor'
-import { Manager } from 'react-popper'
 import pick from 'lodash/pick'
 import pickBy from 'lodash/pickBy'
 import identity from 'lodash/identity'
 import getNested from 'lodash/get'
-import produce from 'immer'
 import { parseCommand, updateTree, removeTemporaryState } from 'vtv'
-import { getTheme } from '../../themes'
 import runCommand from '../../command-runner'
-import { MemoryStore, LocalStorageStore } from '../../store'
-import ConsoleWorkspace from '../../services/workspace/ConsoleWorkspace'
 import Message from '../messages/Message'
 import ChannelInput from './ChannelInput'
 import Nav from '../Nav'
-import Head from '../Head'
 
 class MessageList extends PureComponent {
   render() {
@@ -86,7 +80,7 @@ class MessageList extends PureComponent {
   }
 }
 
-class ChannelView extends PureComponent {
+export default class ChannelView extends PureComponent {
   constructor(props) {
     super(props)
     this.state = {
@@ -390,115 +384,3 @@ class ChannelView extends PureComponent {
     )
   }
 }
-
-class ChannelViewWrapper extends PureComponent {
-  state = {
-    channel: null,
-    theme: null,
-  }
-
-  constructor(props) {
-    super(props)
-    if (this.props.storageType === 'localStorage') {
-      this._store = new LocalStorageStore()
-    } else {
-      this._store = new MemoryStore()
-    }
-    this._store.load()
-    this.state.theme = this._store.theme
-  }
-
-  get store() {
-    return this._store
-  }
-
-  async loadChannel() {
-    const workspace = ConsoleWorkspace.getWorkspace()
-    const channel = await workspace.getChannel('main')
-    this.setState({ channel })
-  }
-
-  get channel() {
-    if (this.props.channel) {
-      return this.props.channel
-    } else if (this.state.channel) {
-      return this.state.channel
-    } else {
-      this.loadChannel()
-    }
-  }
-
-  render() {
-    const { onThemeChange, store, channel, storageType, ...props } = this.props
-    const themeName = this.state.theme
-    const theme = getTheme(themeName)
-    return (
-      <>
-        <ChannelView
-          theme={theme}
-          store={store || this.store}
-          channel={channel || this.channel}
-          onThemeChange={theme => this.setState({ theme })}
-          {...props}
-        />
-        <Head title="Resources.co" theme={theme} />
-        <style jsx global>{`
-          html,
-          body,
-          body > div {
-            margin: 0;
-            padding: 0;
-            height: 100%;
-            box-sizing: border-box;
-          }
-
-          * {
-            margin: 0;
-            padding: 0;
-            font-family: -apple-system, BlinkMacSystemFont, Avenir Next, Avenir,
-              Helvetica, sans-serif;
-          }
-
-          *,
-          *::before,
-          *::after {
-            box-sizing: border-box;
-          }
-
-          html,
-          body {
-            background-color: ${theme.background};
-          }
-
-          html,
-          body,
-          textarea,
-          svg,
-          button {
-            color: ${theme.foreground};
-          }
-
-          a {
-            color: ${theme.linkColor};
-            text-decoration: none;
-          }
-
-          a:hover {
-            text-decoration: underline;
-          }
-
-          ::selection {
-            color: ${theme.selectionColor};
-            background: ${theme.selectionBackground};
-          }
-
-          #__next-prerender-indicator {
-            display: none;
-          }
-        `}</style>
-      </>
-    )
-  }
-}
-
-export default ChannelViewWrapper
