+++ pages/embed/[...path].js
@@ -1,3 +1,12 @@
-export default function Embed({ path }) {
-  return <p>embed {path}</p>
+import dynamic from 'next/dynamic'
+import EmbedPage from 'console/components/channel/EmbedPage'
+
+const CodeMirror = dynamic(() => import('components/CodeMirror'), {
+  ssr: false,
+})
+
+function Embed({ path }) {
+  return <EmbedPage storageType="localStorage" path={path} />
 }
+
+export default dynamic(() => Promise.resolve(Embed), { ssr: false })
